add login and logout

+++ Core/Validator.php
@@ -4,7 +4,7 @@ namespace Core;
 
 readonly class Validator 
 {
-	public static function string(string $value,int $min = 1,int $max = INF): bool
+	public static function string(string $value,int $min = 1,float $max = INF): bool
 	{
 		$value = trim($value);
